Update evaluate_model.r
Vars + IRF does some weird loading of global state. What happens is that it tries to load the `lag` variable from global state which is not defined, but the `lag` function is. Hence it loads that one and everything crashes. Now we actually look for the `p` variable which should be stored globally. This is probably a bug in vars.